+++ apps/admin_audit/appinfo/info.xml
@@ -6,6 +6,7 @@
 	<licence>AGPL</licence>
 	<author>Nextcloud</author>
 	<version>1.3.0</version>
+	<namespace>AdminAudit</namespace>
 	<dependencies>
 		<nextcloud min-version="13" max-version="13" />
 	</dependencies>
