+++ apps/admin_audit/lib/Actions/Action.php
@@ -20,7 +20,8 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  *
  */
-namespace OCA\Admin_Audit\Actions;
+
+namespace OCA\AdminAudit\Actions;
 
 use OCP\ILogger;
