+++ apps/admin_audit/lib/Actions/Files.php
@@ -20,12 +20,13 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  *
  */
-namespace OCA\Admin_Audit\Actions;
+
+namespace OCA\AdminAudit\Actions;
 
 /**
  * Class Files logs the actions to files
  *
- * @package OCA\Admin_Audit\Actions
+ * @package OCA\AdminAudit\Actions
  */
 class Files extends Action {
 	/**
