+++ apps/admin_audit/lib/Actions/Sharing.php
@@ -20,13 +20,16 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  *
  */
-namespace OCA\Admin_Audit\Actions;
+
+namespace OCA\AdminAudit\Actions;
+
+
 use OCP\Share;
 
 /**
  * Class Sharing logs the sharing actions
  *
- * @package OCA\Admin_Audit\Actions
+ * @package OCA\AdminAudit\Actions
  */
 class Sharing extends Action {
 	/**
